# server/api: drop indent, add streaming for Json API (PCP)